+++ lib/whois/parsers/whois.donuts.co.rb
@@ -20,40 +20,36 @@ module Whois
     #
     class WhoisDonutsCo < BaseIcannCompliant
       self.scanner = Scanners::BaseIcannCompliant, {
-          pattern_available: /^Domain not found\.\n/
+          pattern_available: /^Domain not found\./
       }
 
 
-      property_supported :domain_id do
-        node('Domain ID')
-      end
-
-
       property_supported :expires_on do
-        node('Registry Expiry Date') do |value|
+        node("Registry Expiry Date") do |value|
           parse_time(value)
         end
       end
 
 
       property_supported :registrar do
-        return unless node('Sponsoring Registrar')
-        Parser::Registrar.new(
-            id:           node('Sponsoring Registrar IANA ID'),
-            name:         node('Sponsoring Registrar'),
-            organization: node('Sponsoring Registrar')
-        )
+        return unless node("Registrar")
+        Parser::Registrar.new({
+            id:           node("Registrar IANA ID"),
+            name:         node("Registrar"),
+            organization: node("Registrar"),
+            url:          node("Registrar URL"),
+        })
       end
 
 
       private
-
-      def build_contact(element, type)
-        if (contact = super)
-          contact.id = node("#{element} ID")
-        end
-        contact
-      end
+      #
+      # def build_contact(element, type)
+      #   if (contact = super)
+      #     contact.id = node("#{element} ID")
+      #   end
+      #   contact
+      # end
 
     end