+++ spec/whois/parsers/responses/ccwhois.verisign-grs.com/cc/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Parsers::CcwhoisVerisignGrsCom, "status_registered.expected" do
   end
   describe "#domain_id" do
     it do
-      expect(subject.domain_id).to eq("86420657_DOMAIN_CC-VRSN")
+      expect(subject.domain_id).to eq("86420657")
     end
   end
   describe "#status" do
@@ -54,19 +54,19 @@ describe Whois::Parsers::CcwhoisVerisignGrsCom, "status_registered.expected" do
   describe "#created_on" do
     it do
       expect(subject.created_on).to be_a(Time)
-      expect(subject.created_on).to eq(Time.parse("1999-06-07 04:00:00 UTC"))
+      expect(subject.created_on).to eq(Time.parse("1999-06-07 00:00:00 UTC"))
     end
   end
   describe "#updated_on" do
     it do
       expect(subject.updated_on).to be_a(Time)
-      expect(subject.updated_on).to eq(Time.parse("2017-05-06 09:28:40 UTC"))
+      expect(subject.updated_on).to eq(Time.parse("2013-05-06 05:17:44 UTC"))
     end
   end
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2018-06-07 04:00:00 UTC"))
+      expect(subject.expires_on).to eq(Time.parse("2014-06-07 00:00:00 UTC"))
     end
   end
   describe "#registrar" do