+++ spec/whois/parsers/responses/whois.donuts.co/bike/status_registered_spec.rb
@@ -28,7 +28,7 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
   end
   describe "#domain_id" do
     it do
-      expect(subject.domain_id).to eq("C52CECC9AF044831A7335E8A0ECBC349-D")
+      expect(subject.domain_id).to eq("e25432d5c94440c4a8ca0e5ecbc13904-DONUTS")
     end
   end
   describe "#status" do
@@ -49,28 +49,28 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
   describe "#created_on" do
     it do
       expect(subject.created_on).to be_a(Time)
-      expect(subject.created_on).to eq(Time.parse("2014-02-21 22:55:07 UTC"))
+      expect(subject.created_on).to eq(Time.parse("2016-02-26 16:49:10 UTC"))
     end
   end
   describe "#updated_on" do
     it do
       expect(subject.updated_on).to be_a(Time)
-      expect(subject.updated_on).to eq(Time.parse("2014-02-21 22:55:08 UTC"))
+      expect(subject.updated_on).to eq(Time.parse("2017-04-12 16:49:41 UTC"))
     end
   end
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2015-02-21 22:55:07 UTC"))
+      expect(subject.expires_on).to eq(Time.parse("2018-02-26 16:49:10 UTC"))
     end
   end
   describe "#registrar" do
     it do
       expect(subject.registrar).to be_a(Whois::Parser::Registrar)
-      expect(subject.registrar.id).to eq("48")
-      expect(subject.registrar.name).to eq("Enom, Inc.")
-      expect(subject.registrar.organization).to eq("Enom, Inc.")
-      expect(subject.registrar.url).to eq(nil)
+      expect(subject.registrar.id).to eq("146")
+      expect(subject.registrar.name).to eq("GoDaddy.com, LLC")
+      expect(subject.registrar.organization).to eq("GoDaddy.com, LLC")
+      expect(subject.registrar.url).to eq("http://www.godaddy.com/domains/search.aspx?ci=8990")
     end
   end
   describe "#registrant_contacts" do
@@ -79,18 +79,18 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
       expect(subject.registrant_contacts.size).to eq(1)
       expect(subject.registrant_contacts[0]).to be_a(Whois::Parser::Contact)
       expect(subject.registrant_contacts[0].type).to eq(Whois::Parser::Contact::TYPE_REGISTRANT)
-      expect(subject.registrant_contacts[0].id).to eq("8ff85c48fbd456f1")
-      expect(subject.registrant_contacts[0].name).to eq("whoisguard protected")
-      expect(subject.registrant_contacts[0].organization).to eq("WhoisGuard, Inc.")
-      expect(subject.registrant_contacts[0].address).to eq("P.O. Box 0823-03411")
-      expect(subject.registrant_contacts[0].city).to eq("Panama")
-      expect(subject.registrant_contacts[0].zip).to eq("00000")
-      expect(subject.registrant_contacts[0].state).to eq("Panama")
+      expect(subject.registrant_contacts[0].id).to eq("2a94fd50b2ca42c685828dfa8c07e23d-DONUTS")
+      expect(subject.registrant_contacts[0].name).to eq("Marko Matenda")
+      expect(subject.registrant_contacts[0].organization).to eq("")
+      expect(subject.registrant_contacts[0].address).to eq("Ante Starcevica 9.")
+      expect(subject.registrant_contacts[0].city).to eq("Bjelovar")
+      expect(subject.registrant_contacts[0].zip).to eq("43000")
+      expect(subject.registrant_contacts[0].state).to eq("Croatia")
       expect(subject.registrant_contacts[0].country).to eq(nil)
-      expect(subject.registrant_contacts[0].country_code).to eq("PA")
-      expect(subject.registrant_contacts[0].phone).to eq("+507.8365503")
-      expect(subject.registrant_contacts[0].fax).to eq("+51.17057182")
-      expect(subject.registrant_contacts[0].email).to eq("legal@whoisguard.com")
+      expect(subject.registrant_contacts[0].country_code).to eq("HR")
+      expect(subject.registrant_contacts[0].phone).to eq("+385.916283632")
+      expect(subject.registrant_contacts[0].fax).to eq("")
+      expect(subject.registrant_contacts[0].email).to eq("marko.matenda@gmail.com")
       expect(subject.registrant_contacts[0].created_on).to eq(nil)
       expect(subject.registrant_contacts[0].updated_on).to eq(nil)
     end
@@ -101,18 +101,18 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
       expect(subject.admin_contacts.size).to eq(1)
       expect(subject.admin_contacts[0]).to be_a(Whois::Parser::Contact)
       expect(subject.admin_contacts[0].type).to eq(Whois::Parser::Contact::TYPE_ADMINISTRATIVE)
-      expect(subject.admin_contacts[0].id).to eq("8ff85c48fbd456f1")
-      expect(subject.admin_contacts[0].name).to eq("whoisguard protected")
-      expect(subject.admin_contacts[0].organization).to eq("WhoisGuard, Inc.")
-      expect(subject.admin_contacts[0].address).to eq("P.O. Box 0823-03411")
-      expect(subject.admin_contacts[0].city).to eq("Panama")
-      expect(subject.admin_contacts[0].zip).to eq("00000")
-      expect(subject.admin_contacts[0].state).to eq("Panama")
+      expect(subject.admin_contacts[0].id).to eq("a627ad7dc57343858c4397b9e3f9a530-DONUTS")
+      expect(subject.admin_contacts[0].name).to eq("Marko Matenda")
+      expect(subject.admin_contacts[0].organization).to eq("")
+      expect(subject.admin_contacts[0].address).to eq("Ante Starcevica 9.")
+      expect(subject.admin_contacts[0].city).to eq("Bjelovar")
+      expect(subject.admin_contacts[0].zip).to eq("43000")
+      expect(subject.admin_contacts[0].state).to eq("Croatia")
       expect(subject.admin_contacts[0].country).to eq(nil)
-      expect(subject.admin_contacts[0].country_code).to eq("PA")
-      expect(subject.admin_contacts[0].phone).to eq("+507.8365503")
-      expect(subject.admin_contacts[0].fax).to eq("+51.17057182")
-      expect(subject.admin_contacts[0].email).to eq("legal@whoisguard.com")
+      expect(subject.admin_contacts[0].country_code).to eq("HR")
+      expect(subject.admin_contacts[0].phone).to eq("+385.916283632")
+      expect(subject.admin_contacts[0].fax).to eq("")
+      expect(subject.admin_contacts[0].email).to eq("marko.matenda@gmail.com")
       expect(subject.admin_contacts[0].created_on).to eq(nil)
       expect(subject.admin_contacts[0].updated_on).to eq(nil)
     end
@@ -123,18 +123,18 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
       expect(subject.technical_contacts.size).to eq(1)
       expect(subject.technical_contacts[0]).to be_a(Whois::Parser::Contact)
       expect(subject.technical_contacts[0].type).to eq(Whois::Parser::Contact::TYPE_TECHNICAL)
-      expect(subject.technical_contacts[0].id).to eq("8ff85c48fbd456f1")
-      expect(subject.technical_contacts[0].name).to eq("whoisguard protected")
-      expect(subject.technical_contacts[0].organization).to eq("WhoisGuard, Inc.")
-      expect(subject.technical_contacts[0].address).to eq("P.O. Box 0823-03411")
-      expect(subject.technical_contacts[0].city).to eq("Panama")
-      expect(subject.technical_contacts[0].zip).to eq("00000")
-      expect(subject.technical_contacts[0].state).to eq("Panama")
+      expect(subject.technical_contacts[0].id).to eq("08094e7dd78143d6b83338c5c59a8160-DONUTS")
+      expect(subject.technical_contacts[0].name).to eq("Marko Matenda")
+      expect(subject.technical_contacts[0].organization).to eq("")
+      expect(subject.technical_contacts[0].address).to eq("Ante Starcevica 9.")
+      expect(subject.technical_contacts[0].city).to eq("Bjelovar")
+      expect(subject.technical_contacts[0].zip).to eq("43000")
+      expect(subject.technical_contacts[0].state).to eq("Croatia")
       expect(subject.technical_contacts[0].country).to eq(nil)
-      expect(subject.technical_contacts[0].country_code).to eq("PA")
-      expect(subject.technical_contacts[0].phone).to eq("+507.8365503")
-      expect(subject.technical_contacts[0].fax).to eq("+51.17057182")
-      expect(subject.technical_contacts[0].email).to eq("legal@whoisguard.com")
+      expect(subject.technical_contacts[0].country_code).to eq("HR")
+      expect(subject.technical_contacts[0].phone).to eq("+385.916283632")
+      expect(subject.technical_contacts[0].fax).to eq("")
+      expect(subject.technical_contacts[0].email).to eq("marko.matenda@gmail.com")
       expect(subject.technical_contacts[0].created_on).to eq(nil)
       expect(subject.technical_contacts[0].updated_on).to eq(nil)
     end
@@ -142,27 +142,15 @@ describe Whois::Parsers::WhoisDonutsCo, "status_registered.expected" do
   describe "#nameservers" do
     it do
       expect(subject.nameservers).to be_a(Array)
-      expect(subject.nameservers.size).to eq(5)
+      expect(subject.nameservers.size).to eq(2)
       expect(subject.nameservers[0]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[0].name).to eq("dns5.registrar-servers.com")
+      expect(subject.nameservers[0].name).to eq("ns68.domaincontrol.com")
       expect(subject.nameservers[0].ipv4).to eq(nil)
       expect(subject.nameservers[0].ipv6).to eq(nil)
       expect(subject.nameservers[1]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[1].name).to eq("dns3.registrar-servers.com")
+      expect(subject.nameservers[1].name).to eq("ns67.domaincontrol.com")
       expect(subject.nameservers[1].ipv4).to eq(nil)
       expect(subject.nameservers[1].ipv6).to eq(nil)
-      expect(subject.nameservers[2]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[2].name).to eq("dns2.registrar-servers.com")
-      expect(subject.nameservers[2].ipv4).to eq(nil)
-      expect(subject.nameservers[2].ipv6).to eq(nil)
-      expect(subject.nameservers[3]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[3].name).to eq("dns1.registrar-servers.com")
-      expect(subject.nameservers[3].ipv4).to eq(nil)
-      expect(subject.nameservers[3].ipv6).to eq(nil)
-      expect(subject.nameservers[4]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[4].name).to eq("dns4.registrar-servers.com")
-      expect(subject.nameservers[4].ipv4).to eq(nil)
-      expect(subject.nameservers[4].ipv6).to eq(nil)
     end
   end
 end