+++ spec/whois/parsers/responses/whois.verisign-grs.com/net/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Parsers::WhoisVerisignGrsCom, "status_registered.expected" do
   end
   describe "#domain_id" do
     it do
-      expect(subject.domain_id).to eq("4802712_DOMAIN_NET-VRSN")
+      expect(subject.domain_id).to eq(nil)
     end
   end
   describe "#status" do
@@ -54,26 +54,26 @@ describe Whois::Parsers::WhoisVerisignGrsCom, "status_registered.expected" do
   describe "#created_on" do
     it do
       expect(subject.created_on).to be_a(Time)
-      expect(subject.created_on).to eq(Time.parse("1999-03-15 05:00:00 UTC"))
+      expect(subject.created_on).to eq(Time.parse("1999-03-15"))
     end
   end
   describe "#updated_on" do
     it do
       expect(subject.updated_on).to be_a(Time)
-      expect(subject.updated_on).to eq(Time.parse("2017-02-11 10:56:37 UTC"))
+      expect(subject.updated_on).to eq(Time.parse("2014-02-11"))
     end
   end
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2018-03-15 04:00:00 UTC"))
+      expect(subject.expires_on).to eq(Time.parse("2015-03-15"))
     end
   end
   describe "#registrar" do
     it do
       expect(subject.registrar).to be_a(Whois::Parser::Registrar)
-      expect(subject.registrar.id).to eq("292")
-      expect(subject.registrar.name).to eq("MarkMonitor Inc.")
+      expect(subject.registrar.id).to eq(nil)
+      expect(subject.registrar.name).to eq("MARKMONITOR INC.")
       expect(subject.registrar.organization).to eq(nil)
       expect(subject.registrar.url).to eq("http://www.markmonitor.com")
     end