+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 #### master
 
-- FIXED: Updated .BR parser to the new response (GH-51). [Thanks @otaviojr]
+- CHANGED: Updated .BR parser to the new response (GH-51). [Thanks @otaviojr]
+
+- CHANGED: Add support for :expires_on to base_nic_fr (GH-54). [Thanks @yastupin]
 
 
 #### Release 1.0
@@ -26,7 +28,7 @@ Initial import from the `whois` library.
 
 - NEW: Detect reserved .INFO domains (whois/GH-468).
 
-- FIXED: whois.audns.net.au removed the registrar ID field (GH-20, GH-21). Thanks [@afoster]
+- CHANGED: whois.audns.net.au removed the registrar ID field (GH-20, GH-21). Thanks [@afoster]
 
 - CHANGED: Updated .JOBS from obswhois.verisign-grs.com to whois.nic.jobs (GH-23).